requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
opencv-python>=4.8.0
numpy>=1.26.0
pillow>=10.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.0
aiofiles>=23.2.0
sqlalchemy>=2.0.0
httpx>=0.25.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
PyJWT>=2.8.0
